refactor: move display logic and reviewer extraction to PullRequest struct
